+++ next/src/app/profile/page.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+export default function SearchPage() {
+  return (
+    <>
+      <h1 className="text-2xl font-bold mb-4">プロフィール</h1>
+      <p>ここにプロフィール情報を表示します。</p>
+    </>
+  );
+}
