+++ next/src/components/Header.tsx
@@ -61,13 +61,13 @@ export const Header = () => {
                 // ログインしていない場合
                 <>
                   <Link
-                    href="/sign-in"
+                    href="/auth/sign-in"
                     className="text-gray-600 hover:text-gray-900"
                   >
                     サインイン
                   </Link>
                   <Link
-                    href="/sign-up"
+                    href="/auth/sign-up"
                     className="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600"
                   >
                     サインアップ
